added open experience live api with tests
