+++ pyaffalddk/municipalities.py
@@ -119,7 +119,9 @@ MUNICIPALITIES_LIST = {
     "Favrskov": ["affaldonline", "ZGZmY2M1YjYtYjllZS00NzhkLTgyZTItMDMwMTIzNDg1Zjdl"],
     "Faxe": ["perfectwaste"],
     "Fredensborg": ["renoweb"],
+    "Fredericia": ["openexplive", "3YWh0MjlpbDh1djNiM25hZA=="],
     "Frederikssund": ["perfectwaste"],
+    "Frederiksberg": ["openexplive", '6RjFCNUdtUGNxRU1xMGxFWA=='],
     "Furesø": ["vestfor"],
     "Gentofte": ["renoweb"],
     "Gladsaxe": ["perfectwaste"],
@@ -153,7 +155,7 @@ MUNICIPALITIES_LIST = {
     "Nordfyns": ['openexp', 'reno.nordfynskommune.dk/app/AppService/AppService/'],
     "Odense": ["odense"],
     "Odsherred": ["perfectwaste"],
-    "Randers": ["renoweb"],
+    "Randers": ["renoweb"],  # also openexplive 4UGdDU2dhdDFqTHZsV0NQSQ==
     "Rebild": ["affaldonline", "Y2IzZGRjOGMtOTAwYS00M2NlLWFjODgtZWM3NTg3ZGI0ZGIz"],
     "Ringkøbing-Skjern": ["perfectwaste"],
     "Roskilde": ["perfectwaste"],
@@ -165,6 +167,7 @@ MUNICIPALITIES_LIST = {
     "Stevns": ["perfectwaste"],
     "Svendborg": ["renoweb"],
     "Sønderborg": ["renoweb"],
+    "Thy": ["openexplive", "5NmkzUGpvZlRaMzdqZzBEQw=="],
     "Tårnby": ["perfectwaste"],
     "Vallensbæk": ["vestfor"],
     "Varde": ["renoweb"],
