+++ tests/data/openexplive.data
@@ -1,0 +1,1 @@
+[{"arrangement_id": 145570, "fraction": {"id": 34, "name": "Restaffald"}, "is_domestic": true, "can_create_missed_collection": true, "can_order_collection": false, "last_past_date": {"date": "2025-05-13", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, "upcoming_dates": [{"date": "2025-05-27", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-06-10", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-06-24", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-07-08", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-07-22", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-08-05", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-08-19", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}], "sorting_types": []}, {"arrangement_id": 140199, "fraction": {"id": 7, "name": "Haveaffald"}, "is_domestic": false, "can_create_missed_collection": true, "can_order_collection": false, "last_past_date": {"date": "2025-05-08", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, "upcoming_dates": [{"date": "2025-05-29", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-06-19", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-07-10", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-07-31", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-08-21", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}], "sorting_types": []}, {"arrangement_id": 26422, "fraction": {"id": 20, "name": "Stort Elektronik"}, "is_domestic": false, "can_create_missed_collection": true, "can_order_collection": false, "last_past_date": {"date": "2025-03-06", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, "upcoming_dates": [{"date": "2025-05-29", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}], "sorting_types": []}, {"arrangement_id": 26428, "fraction": {"id": 6, "name": "Storskrald"}, "is_domestic": false, "can_create_missed_collection": true, "can_order_collection": false, "last_past_date": {"date": "2025-03-06", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, "upcoming_dates": [{"date": "2025-05-29", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}], "sorting_types": []}, {"arrangement_id": 145571, "fraction": {"id": 15, "name": "Madaffald"}, "is_domestic": false, "can_create_missed_collection": true, "can_order_collection": false, "last_past_date": {"date": "2025-05-13", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, "upcoming_dates": [{"date": "2025-05-27", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-06-10", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-06-24", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-07-08", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-07-22", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-08-05", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-08-19", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}], "sorting_types": []}, {"arrangement_id": 140202, "fraction": {"id": 4, "name": "Pap"}, "is_domestic": false, "can_create_missed_collection": true, "can_order_collection": false, "last_past_date": {"date": "2025-05-12", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, "upcoming_dates": [{"date": "2025-06-02", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-06-23", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-07-14", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-08-04", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}], "sorting_types": []}, {"arrangement_id": 140201, "fraction": {"id": 50, "name": "Metal/papir"}, "is_domestic": false, "can_create_missed_collection": true, "can_order_collection": false, "last_past_date": {"date": "2025-05-12", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, "upcoming_dates": [{"date": "2025-07-07", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}], "sorting_types": []}, {"arrangement_id": 140198, "fraction": {"id": 51, "name": "Plast/ Mad- og drikkekartoner"}, "is_domestic": false, "can_create_missed_collection": true, "can_order_collection": false, "last_past_date": {"date": "2025-05-08", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, "upcoming_dates": [{"date": "2025-05-29", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-06-19", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-07-10", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-07-31", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-08-21", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}], "sorting_types": []}, {"arrangement_id": 162694, "fraction": {"id": 53, "name": "Tekstilaffald"}, "is_domestic": false, "can_create_missed_collection": true, "can_order_collection": false, "last_past_date": {"date": "2025-05-12", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, "upcoming_dates": [{"date": "2025-06-02", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-06-23", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-07-14", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}, {"date": "2025-08-04", "is_special": false, "special_note": null, "complaint_date": null, "complaints": []}], "sorting_types": []}]
